feat(mobile+web): functional gaps — settings stack, onboarding bootstrap, preferred-currency plumbing
- Move Settings from tabs to a stack route so it gets a back affordance
- Onboarding now seeds the local SQLite profile/account rows via
  `bootstrapOnboardingLocally`, so the first dashboard render after signup
  doesn't wait on a full sync cycle
- Wire `preferred_currency` through 7 Root screens + dashboard summary
  (was hard-coded "COP"), cached in-memory and invalidated on profile pull
  so filter changes don't re-hit SQLite
- Webapp onboarding persists a `mobile_dashboard_config` default so the
  native app has a layout on first launch
- `getDashboardHeroData` reads onboarding-declared income so the pay-cycle
  window is configured even before the user adds a recurring income;
  wrapped in `"use cache"` + `cacheTag("profile")` to keep the hero off
  the live DB path
- Unify `MobileDashboardLayout` as an alias of `@zeta/shared`'s
  `DashboardLayout` so the two platforms can't silently drift
- Fix stack-route padding on Settings (`insets.bottom` instead of tab-bar
  clearance)
