+++ mobile/app/(tabs)/_layout.tsx
@@ -17,7 +17,6 @@ export default function TabLayout() {
       <Tabs.Screen name="import" options={{ href: null }} />
       <Tabs.Screen name="accounts" options={{ href: null }} />
       <Tabs.Screen name="budgets" options={{ href: null }} />
-      <Tabs.Screen name="settings" options={{ href: null }} />
     </Tabs>
   );
 }
